+++ src/utils/networking.cpp
@@ -8,12 +8,11 @@ std::string loopbackInterfaceName()
   return "lo";
 #elif defined(__APPLE__) || defined(BSD)
   return "lo0";
-#elif defined(__WIN32)
+#elif defined(_WIN32)
   // Not required as we directly use the 127.0.0.1 under Windows
   return "";
 #else
-#warning "Your target architecture does not define a loopback interface. Please consider reporting this to the preCICE developers."
-  return "";
+#error "There is no loopback device defined for your OS. Please open an issue with a suggestion at https://github.com/precice/precice/issues"
 #endif
 }